% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivo.R
\docType{data}
\name{font_ar}
\alias{font_ar}
\alias{font_ar_bold}
\alias{font_ar_italic}
\alias{font_ar_sb}
\alias{font_ar_sb_italic}
\title{Archivo SemiCondensed font name R variable aliases}
\format{
length 1 character vector

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
font_ar

font_ar_bold

font_ar_italic

font_ar_sb

font_ar_sb_italic
}
\description{
\code{font_ar} == "\verb{Archivo SemiCondensed}"

\code{font_ar_bold} == "\verb{Archivo SemiCondensed Bold}"

\code{font_ar_italic} == "\verb{Archivo SemiCondensed Italic}"

\code{font_ar_sb} == "\verb{Archivo SemiCondensed SemiBold}"

\code{font_ar_sb_italic} == "\verb{Archivo SemiCondensed SemiBold Italic}"
}
\keyword{datasets}
