% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit12.R
\docType{data}
\name{bit12}
\alias{bit12}
\title{Bit12 Color Palette}
\format{
A named character vector of hex color codes
}
\usage{
bit12
}
\description{
A vector of 12 colors in the Bit12 color scheme
}
\keyword{datasets}
