% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit12.R
\docType{data}
\name{bit12_extended}
\alias{bit12_extended}
\title{Extended Bit12 Color Palettes}
\format{
A list of character vectors containing hex color codes
}
\usage{
bit12_extended
}
\description{
A list of 12 color palettes, each containing 9 shades for continuous color interpolation
}
\keyword{datasets}
