% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colors.R
\name{create_continuous_scale}
\alias{create_continuous_scale}
\title{Color Utilities for hrbrthemes}
\usage{
create_continuous_scale(colors)
}
\arguments{
\item{colors}{Vector of hex color codes}
}
\value{
A color ramp palette function
}
\description{
Internal helper functions for color palette operations.
}
\keyword{Color}
\keyword{Continuous}
\keyword{Create}
\keyword{Internal}
\keyword{Scale}
\keyword{a}
\keyword{color}
\keyword{colors}
\keyword{continuous}
\keyword{from}
\keyword{function}
\keyword{generate}
\keyword{internal}
\keyword{of}
\keyword{scales}
\keyword{to}
\keyword{vector}
