% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ-sans.R
\docType{data}
\name{font_es}
\alias{font_es}
\alias{font_es_bold}
\alias{font_es_light}
\title{Econ Sans Condensed font name R variable aliases}
\format{length 1 character vector}
\usage{
font_es

font_es_bold

font_es_light
}
\description{
\code{font_es} == "\code{EconSansCndLig}"

\code{font_es_bold} == "\code{EconSansCndBol}"

\code{font_es_light} == "\code{EconSansCndLig}"
}
\note{
\code{font_es_bold} (a.k.a. "\code{EconSansCndBol}") is not available on
Windows and will throw a warning if used in plots.

\code{font_es_light} (a.k.a. "\code{EconSansCndLig}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
