% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki.R
\docType{data}
\name{flexoki_dark}
\alias{flexoki_dark}
\title{Flexoki Dark Color Palette}
\format{
A named character vector of hex color codes
}
\usage{
flexoki_dark
}
\description{
A vector of 8 colors in the dark variant of the Flexoki color scheme
}
\keyword{datasets}
