% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki.R
\docType{data}
\name{flexoki_extended}
\alias{flexoki_extended}
\title{Extended Flexoki Color Palettes}
\format{
A list of character vectors containing hex color codes
}
\usage{
flexoki_extended
}
\description{
A list of 8 color palettes, each containing 13 shades for continuous color interpolation
}
\keyword{datasets}
