% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki.R
\docType{data}
\name{flexoki_light}
\alias{flexoki_light}
\title{Flexoki Light Color Palette}
\format{
A named character vector of hex color codes
}
\usage{
flexoki_light
}
\description{
A vector of 8 colors in the light variant of the Flexoki color scheme
}
\keyword{datasets}
