% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush-ticks.R
\name{flush_ticks}
\alias{flush_ticks}
\title{Makes axis text labels flush on the ends}
\usage{
flush_ticks(gg, flush = "XY", plot = TRUE, cat = TRUE)
}
\arguments{
\item{gg}{ggplot2 plot object}

\item{flush}{either "\code{X}" or "\code{Y}" or "\code{XY}" to flush individual or both axes. Default: both.}

\item{plot}{if \code{FALSE} then the ggplot object will be returned \emph{invisibly}}

\item{cat}{if \code{TRUE} then display \code{theme()} statements and copy them to the clipboard}
}
\value{
ggplot2 object with \code{theme()} elements added
}
\description{
A covenience function intended for basic, fixed-scale plots only (i.e. does not handle
free scales in facets).\cr
\cr
You need to pass in a ggplot2 object to this function. It can't be \code{+}'d in a chain
of geoms, coords, scales, themes, etc. It also builds the plot (but does not display
it) so if the plt takes a while (i.e. has lots of data or transforms) this will also
take a while.
}
\details{
If \code{clipr} is installed and \code{cat} is \code{TRUE} then the \code{theme(...)} statements will
be copied to the clipboard as well as displayed (via \code{cat()}) in the console.
}
\note{
Intended for basic, fixed-scale plots only (i.e. does not handle free scales in facets).
Also, the clipboard operations will only occur if \code{clipr} is installed.
}
