% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\docType{data}
\name{ft_cols}
\alias{ft_cols}
\alias{ft_text_col}
\title{FT color palette}
\format{
An object of class \code{list} of length 9.

An object of class \code{character} of length 1.
}
\usage{
ft_cols

ft_text_col
}
\description{
FT color palette
}
\note{
don't forget you can use \code{\link[scales:alpha]{scales::alpha()}} with these colors
}
\keyword{datasets}
