% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{ft_geom_defaults}
\alias{ft_geom_defaults}
\title{Change geom defaults from black to custom lights for the FT theme}
\usage{
ft_geom_defaults()
}
\description{
Change geom defaults from black to custom lights for the FT theme
}
