% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldmansans.R
\docType{data}
\name{font_gs}
\alias{font_gs}
\title{Goldman Sans font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_gs
}
\description{
\code{font_gs} == "\verb{Goldman Sans Condensed}"
}
\keyword{datasets}
