% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrbrthemes-package.R
\name{hrbrthemes-exports}
\alias{\%>\%}
\alias{hrbrthemes-exports}
\title{hrbrthemes exported operators}
\description{
The following functions are imported and then re-exported
from the hrbrthemes package to enable use of the magrittr
pipe operator with no additional library calls
}

