% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrbrthemes-package.R
\docType{package}
\name{hrbrthemes}
\alias{hrbrthemes}
\title{Additional Themes and Theme Components for 'ggplot2'}
\description{
A compilation of extra themes and theme components for 'ggplot2' with an
emphasis on typography.
}
\details{
The core theme: \code{theme_ipsum} ("ipsum" is Latin for "precise") uses Arial Narrow
which should be installed on practically any modern system, so it's "free"-ish.
This font is condensed, has solid default kerning pairs and geometric numbers.
That's what I consider the "font trifecta" must-have for charts. An additional
quality for fonts for charts is that they have a diversity of weights. Arial
Narrow (the one on most systems, anyway) does not have said diversity but this
quality is not (IMO) a "must have".

There is an option \code{hrbrthemes.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
