% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivo.R
\name{import_archivo}
\alias{import_archivo}
\title{Import Archivo SemiCondensed font for use in charts}
\usage{
import_archivo()
}
\description{
Archivo is a grotesque font reminiscent of late nineteenth century.
It is copyright 2019 Omnibus-Type and is licensed under the SIL Open Font License, 1.1;
}
\details{
There is an option \code{hrbrthemes.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
