% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ-sans.R
\name{import_econ_sans}
\alias{import_econ_sans}
\title{Import Roboto Condensed font for use in charts}
\usage{
import_econ_sans()
}
\description{
Roboto Condensed is a trademark of Google.
}
\details{
There is an option \code{hrbrthemes.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
