% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldmansans.R
\name{import_goldman_sans}
\alias{import_goldman_sans}
\title{Import Goldman Sans font for use in charts}
\usage{
import_goldman_sans()
}
\description{
Goldman Sans is a trademark of Goldman Sachs and distributed under the
Goldman Sachs Restricted Font License
}
\details{
There is an option \code{hrbrthemes.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
font directory is displayed after the base import is complete. It is highly
recommended that you install them on your system the same way you would any
other font you wish to use in other programs.
}
