% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{import_inter}
\alias{import_inter}
\title{Import Inter font for use in charts}
\usage{
import_inter()
}
\description{
Inter is Copyright (c) 2016-2024 The Inter Project Authors
}
\details{
There is an option \code{hrbrthemes.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
