% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-sans.R
\name{import_public_sans}
\alias{import_public_sans}
\title{Import Public Sans font for use in charts}
\usage{
import_public_sans()
}
\description{
Public Sans is Copyright 2015 Impallari Type and licensed under the
SIL Open Font License, Version 1.1
}
\details{
There is an option \code{hrbrthemes.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
