% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roboto-condensed.r
\name{import_roboto_condensed}
\alias{import_roboto_condensed}
\title{Import Roboto Condensed font for use in charts}
\usage{
import_roboto_condensed()
}
\description{
Roboto Condensed is a trademark of Google.
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
  font directory is displayed after the base import is complete. It is highly
  recommended that you install them on your system the same way you would any
  other font you wish to use in other programs.
}
