% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\docType{data}
\name{font_inter_thin}
\alias{font_inter_thin}
\alias{font_inter_thin_italic}
\alias{font_inter_medium}
\alias{font_inter_medium_italic}
\alias{font_inter_semibold}
\alias{font_inter_bold}
\alias{font_inter_italic}
\title{Inter font name R variable aliases}
\format{
length 1 character vector

length 1 character vector

length 1 character vector

length 1 character vector

length 1 character vector

length 1 character vector

length 1 character vector
}
\usage{
font_inter_thin

font_inter_thin_italic

font_inter_medium

font_inter_medium_italic

font_inter_semibold

font_inter_bold

font_inter_italic
}
\description{
\code{font_inter_thin} == "\code{Inter-Thin}"

\code{font_inter_thin_italic} == "\code{Inter-ThinItalic}"

\code{font_inter_medium} == "\code{Inter-Medium}"

\code{font_inter_medium_italic} == "\code{Inter-MediumItalic}"

\code{font_inter_semibold} == "\code{Inter-SemiBold}"

\code{font_inter_bold} == "\code{Inter-Bold}"

\code{font_inter_italic} == "\code{Inter-Italic}"
}
\keyword{datasets}
