% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{ipsum_pal}
\alias{ipsum_pal}
\title{A muted, qualitative color palette}
\usage{
ipsum_pal()
}
\description{
A muted, qualitative color palette
}
\examples{
library(scales)
scales::show_col(ipsum_pal()(9))
}

