% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maven.r
\docType{data}
\name{font_mp}
\alias{font_mp}
\title{Maven Pro font name R variable aliases}
\format{length 1 character vector}
\usage{
font_mp
}
\description{
\code{font_mp} = = "\code{Maven Pro}"
}
\keyword{datasets}
