% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roboto-condensed.r
\name{modern_geom_defaults}
\alias{modern_geom_defaults}
\title{Change geom defaults from black to white for the modern theme}
\usage{
modern_geom_defaults()
}
\description{
Change geom defaults from black to white for the modern theme
}
