% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plexsans.R
\docType{data}
\name{font_ps}
\alias{font_ps}
\alias{font_ps_light}
\title{PlexSans font name R variable aliases}
\format{length 1 character vector}
\usage{
font_ps

font_ps_light
}
\description{
\code{font_ps} == "\code{IBMPlexSans}"

\code{font_ps_light} == "\code{IBMPlexSans-Light}"
}
\note{
\code{font_ps_light} (a.k.a. "\code{IBMPlexSans-Light}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
