% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-sans.R
\docType{data}
\name{font_pub}
\alias{font_pub}
\alias{font_pub_bold}
\alias{font_pub_light}
\alias{font_pub_thin}
\title{Public Sans font name R variable aliases}
\format{length 1 character vector}
\usage{
font_pub

font_pub_bold

font_pub_light

font_pub_thin
}
\description{
\code{font_pub} == "\code{Public Sans}"

\code{font_pub_bold} == "\code{Public Sans Bold}"

\code{font_pub_light} == "\code{Public Sans Light}"

\code{font_pub_thin} == "\code{Public Sans Thin}"
}
\note{
\code{font_pub_bold} (a.k.a. "\code{Public Sans Bold}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
