% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roboto-condensed.r
\docType{data}
\name{font_rc}
\alias{font_rc}
\alias{font_rc_light}
\title{Roboto Condensed font name R variable aliases}
\format{
length 1 character vector

An object of class \code{character} of length 1.
}
\usage{
font_rc

font_rc_light
}
\description{
\code{font_rc} == "\verb{Roboto Condensed}"

\code{font_fc_light} == "\verb{Roboto Condensed Light}"
}
\note{
\code{font_rc_light} (a.k.a. "\verb{Roboto Condensed Light}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
