% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit12.R
\name{scale_color_bit12}
\alias{scale_color_bit12}
\title{Discrete Color Scale Using Bit12 Colors}
\usage{
scale_color_bit12(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_color_manual()}
}
\value{
A discrete ggplot2 color scale
}
\description{
Discrete Color Scale Using Bit12 Colors
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point() +
  scale_color_bit12()
}
