% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit12.R
\name{scale_color_bit12_continuous}
\alias{scale_color_bit12_continuous}
\title{Continuous Color Scale Using Bit12 Colors}
\usage{
scale_color_bit12_continuous(palette = "azure", ...)
}
\arguments{
\item{palette}{Name of the color palette to use ("plum", "rose", "coral", "apricot", "lemon", "lime", "mint", "teal", "sky", "azure", "cobalt", "violet")}

\item{...}{Additional arguments passed to scale_color_gradientn()}
}
\value{
A continuous ggplot2 color scale
}
\description{
Continuous Color Scale Using Bit12 Colors
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, color = density)) +
  geom_point() +
  scale_color_bit12_continuous(palette = "azure")
}
