% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit12.R
\name{scale_color_bit12_distiller}
\alias{scale_color_bit12_distiller}
\title{Distiller Color Scale Using Bit12 Colors}
\usage{
scale_color_bit12_distiller(palette = "azure", direction = 1, ...)
}
\arguments{
\item{palette}{Name of the color palette to use ("plum", "rose", "coral", "apricot", "lemon", "lime", "mint", "teal", "sky", "azure", "cobalt", "violet")}

\item{direction}{Sets the direction of the color scale (1 = default, -1 = reversed)}

\item{...}{Additional arguments passed to scale_color_gradientn()}
}
\value{
A sequential ggplot2 color scale
}
\description{
Creates a sequential color gradient based on a selected Bit12 color
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, color = density)) +
  geom_point() +
  scale_color_bit12_distiller(palette = "azure")
}
