% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki.R
\name{scale_color_flexoki_dark}
\alias{scale_color_flexoki_dark}
\title{Discrete Color Scale Using Flexoki Dark Colors}
\usage{
scale_color_flexoki_dark(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_color_manual()}
}
\value{
A discrete ggplot2 color scale
}
\description{
Discrete Color Scale Using Flexoki Dark Colors
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point() +
  scale_color_flexoki_dark()
}
