% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki.R
\name{scale_color_flexoki_dark_distiller}
\alias{scale_color_flexoki_dark_distiller}
\title{Distiller Color Scale Using Flexoki Dark Colors}
\usage{
scale_color_flexoki_dark_distiller(palette = "blue", direction = 1, ...)
}
\arguments{
\item{palette}{Name of the color palette to use ("red", "orange", "yellow", "green", "cyan", "blue", "purple", "magenta")}

\item{direction}{Sets the direction of the color scale (1 = default, -1 = reversed)}

\item{...}{Additional arguments passed to scale_color_gradientn()}
}
\value{
A sequential ggplot2 color scale
}
\description{
Creates a sequential color gradient based on the Flexoki dark color palette
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, color = density)) +
  geom_point() +
  scale_color_flexoki_dark_distiller(palette = "blue")
}
