% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki.R
\name{scale_color_flexoki_light_spectrum}
\alias{scale_color_flexoki_light_spectrum}
\title{Distiller Color Scale Across All Flexoki Light Colors}
\usage{
scale_color_flexoki_light_spectrum(direction = 1, ...)
}
\arguments{
\item{direction}{Sets the direction of the color scale (1 = default, -1 = reversed)}

\item{...}{Additional arguments passed to scale_color_gradientn()}
}
\value{
A sequential ggplot2 color scale
}
\description{
Creates a sequential color gradient using all colors from the Flexoki light palette
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, color = density)) +
  geom_point() +
  scale_color_flexoki_light_spectrum()
}
