% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit12.R
\name{scale_fill_bit12}
\alias{scale_fill_bit12}
\title{Discrete Fill Scale Using Bit12 Colors}
\usage{
scale_fill_bit12(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_fill_manual()}
}
\value{
A discrete ggplot2 fill scale
}
\description{
Discrete Fill Scale Using Bit12 Colors
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_bit12()
}
