% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit12.R
\name{scale_fill_bit12_spectrum}
\alias{scale_fill_bit12_spectrum}
\title{Distiller Fill Scale Across All Bit12 Colors}
\usage{
scale_fill_bit12_spectrum(direction = 1, ...)
}
\arguments{
\item{direction}{Sets the direction of the color scale (1 = default, -1 = reversed)}

\item{...}{Additional arguments passed to scale_fill_gradientn()}
}
\value{
A sequential ggplot2 fill scale
}
\description{
Creates a sequential fill gradient using all colors from the Bit12 palette
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_bit12_spectrum()
}
