% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki.R
\name{scale_fill_flexoki_dark}
\alias{scale_fill_flexoki_dark}
\title{Discrete Fill Scale Using Flexoki Dark Colors}
\usage{
scale_fill_flexoki_dark(...)
}
\arguments{
\item{...}{Additional arguments passed to scale_fill_manual()}
}
\value{
A discrete ggplot2 fill scale
}
\description{
Discrete Fill Scale Using Flexoki Dark Colors
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_flexoki_dark()
}
