% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki.R
\name{scale_fill_flexoki_dark_spectrum}
\alias{scale_fill_flexoki_dark_spectrum}
\title{Distiller Fill Scale Across All Flexoki Dark Colors}
\usage{
scale_fill_flexoki_dark_spectrum(direction = 1, ...)
}
\arguments{
\item{direction}{Sets the direction of the color scale (1 = default, -1 = reversed)}

\item{...}{Additional arguments passed to scale_fill_gradientn()}
}
\value{
A sequential ggplot2 fill scale
}
\description{
Creates a sequential fill gradient using all colors from the Flexoki dark palette
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_flexoki_dark_spectrum()
}
