% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivo.R
\name{theme_ipsum_archivo}
\alias{theme_ipsum_archivo}
\title{A precise & pristine [ggplot2][ggplot2][ggplot2::ggplot2] theme with opinionated defaults and an emphasis on typoghraphy}
\usage{
theme_ipsum_archivo(
  base_family = "Archivo SemiCondensed",
  base_size = 11.5,
  plot_title_family = "Archivo SemiCondensed Bold",
  plot_title_size = 18,
  plot_title_face = "plain",
  plot_title_margin = 10,
  subtitle_family = if (.Platform$OS.type == "windows") "Archivo SemiCondensed" else
    "Archivo SemiCondensed",
  subtitle_size = 13,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = "Archivo SemiCondensed SemiBold",
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = if (.Platform$OS.type == "windows") "Archivo SemiCondensed" else
    "Archivo SemiCondensed",
  caption_size = 9,
  caption_face = "plain",
  caption_margin = 10,
  axis_text_size = 9,
  axis_title_family = base_family,
  axis_title_size = 9,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30),
  grid_col = "#cccccc",
  grid = TRUE,
  axis_col = "#cccccc",
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot tilte family, face, size and margin}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justification one of \verb{[blmcrt]}}

\item{plot_margin}{plot margin (specify with \link[ggplot2:element]{ggplot2::margin})}

\item{grid_col}{grid color}

\item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis_col}{axis color}

\item{axis}{add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}"}

\item{ticks}{ticks if \code{TRUE} add ticks}
}
\description{
You should \code{\link[=import_archivo]{import_archivo()}} first and install the fonts on your
system before trying to use this theme.
}
\details{
There is an option \code{hrbrthemes.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
\section{Why Archivo?}{

Archivo was designed to be used simultaneously in print and digital platforms.
The technical and aesthetic characteristics of the font are both crafted for
high performance typography. It was designed by Héctor Gatti to be used
simultaneously in print and online platforms and supports over 200 world languages
and has tabular numbers. Archivo is a grotesque sans serif typeface family from
Omnibus-Type. It was originally designed for highlights and headlines.
This family is reminiscent of late nineteenth century American typefaces.
}

