% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyhand.R
\docType{data}
\name{font_th}
\alias{font_th}
\title{TinyHand Web font name R variable aliases}
\format{length 1 character vector}
\usage{
font_th
}
\description{
\code{font_th} == "\verb{BF Tiny Hand}"
}
\keyword{datasets}
