% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ipsum.r
\name{update_geom_font_defaults}
\alias{update_geom_font_defaults}
\title{Update matching font defaults for text geoms}
\usage{
update_geom_font_defaults(family = "Arial Narrow", face = "plain",
  size = 3.5)
}
\arguments{
\item{family, face, size}{font family name, face and size}
}
\description{
Updates [ggplot2::geom_label] and [ggplot2::geom_text] font defaults
}

