% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colors.R
\name{validate_direction}
\alias{validate_direction}
\title{Validate Direction Parameter}
\usage{
validate_direction(direction)
}
\arguments{
\item{direction}{The direction parameter to validate}
}
\description{
Checks if direction is 1 or -1 and aborts with a helpful error if not.
}
\keyword{internal}
