% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki.R
\name{create_continuous_scale}
\alias{create_continuous_scale}
\title{Create Continuous Color Scale}
\usage{
create_continuous_scale(colors)
}
\arguments{
\item{colors}{Vector of hex color codes}
}
\value{
A color ramp palette function
}
\description{
Internal function to generate continuous color scales from a vector of colors
}
\keyword{internal}
