% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\docType{data}
\name{ft_cols}
\alias{ft_cols}
\alias{ft_text_col}
\title{FT color palette}
\format{An object of class \code{list} of length 8.}
\usage{
ft_cols

ft_text_col
}
\description{
FT color palette
}
\keyword{datasets}
