% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roboto-condensed.r
\docType{data}
\name{font_rc}
\alias{font_rc}
\alias{font_rc_light}
\title{Roboto Condensed font name R variable aliases}
\format{length 1 character vector}
\usage{
font_rc

font_rc_light
}
\description{
\code{font_rc} == "\code{Roboto Condensed}"

\code{font_fc_light} == "\code{Roboto Condensed Light}"
}
\keyword{datasets}
