% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{scale_colour_ipsum}
\alias{scale_color_ipsum}
\alias{scale_colour_ipsum}
\alias{scale_fill_ipsum}
\title{Discrete color & fill scales based on the ipsum palette}
\usage{
scale_colour_ipsum(...)

scale_color_ipsum(...)

scale_fill_ipsum(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
See [ipsum_pal]().
}

